🎨 Show_cad deduplication (#4379)
* 🎨 Show_cad deduplication

* ✅ Move tests
