+++ bluemira/codes/cadapi/_cadquery/display.py
@@ -148,39 +148,6 @@ class DefaultDisplayOptions:
         self.colour = value
 
 
-def _compute_default_camera(parts: list[apiShape]) -> tuple[tuple, tuple]:
-    """Pick a sensible polyscope camera position based on geometry extents.
-
-    Auto-aligns to the flat axis when the geometry is planar (e.g. the wires
-    produced by ``reactor.show_cad('xz')`` all live at y≈0); falls back to a
-    FreeCAD-style xz side-view for genuinely 3D scenes. Returns
-    ``(camera_position, look_at_target)`` as 3-tuples.
-    """
-    from bluemira.codes import _geometryapi as cadapi  # noqa: PLC0415
-
-    bbs = [cadapi.bounding_box(p) for p in parts]
-    xmin = min(b[0] for b in bbs)
-    ymin = min(b[1] for b in bbs)
-    zmin = min(b[2] for b in bbs)
-    xmax = max(b[3] for b in bbs)
-    ymax = max(b[4] for b in bbs)
-    zmax = max(b[5] for b in bbs)
-    center = ((xmin + xmax) / 2, (ymin + ymax) / 2, (zmin + zmax) / 2)
-    ex, ey, ez = xmax - xmin, ymax - ymin, zmax - zmin
-    m = max(ex, ey, ez) or 1.0
-    d = 1.5 * m  # camera distance — 1.5x the dominant extent gives a comfortable framing
-    flat = 0.01 * m  # near-zero extent threshold (1% of dominant)
-    if ey < flat:  # flat in y → look at xz plane from -y (the dim='xz' case)
-        cam = (center[0], center[1] - d, center[2])
-    elif ex < flat:  # flat in x → look at yz plane from +x
-        cam = (center[0] + d, center[1], center[2])
-    elif ez < flat:  # flat in z → look at xy plane from +z (top-down)
-        cam = (center[0], center[1], center[2] + d)
-    else:  # 3D → xz side-view, mirrors _freecad.api.show_cad's (90,0,0) default
-        cam = (center[0], center[1] - d, center[2])
-    return cam, center
-
-
 def show_cad(
     parts: apiShape | list[apiShape],
     part_options: list[dict],
@@ -207,36 +174,17 @@ def show_cad(
         Coin3D viewer. Mouse navigation (pan/zoom/rotate) is unaffected either
         way.
     """
-    import polyscope as ps  # noqa: PLC0415
-
     from bluemira.codes import _polyscope as ps_backend  # noqa: PLC0415
 
-    parts_list = parts if isinstance(parts, list) else [parts]
+    # Allows separate defaults to polyscope
+    if None in part_options:
+        part_options = [
+            DefaultDisplayOptions() if o is None else o for o in part_options
+        ]
 
-    transparency = "none"
-    for opt in part_options or []:
-        if opt is not None and not np.isclose(opt["transparency"], 0):
-            transparency = "pretty"
-            break
-
-    ps_backend.polyscope_setup(
-        up_direction=kwargs.get("up_direction", "z_up"),
-        fps=kwargs.get("fps", 60),
-        aa=kwargs.get("aa", 1),
-        transparency=transparency,
-        render_passes=kwargs.get("render_passes", 3),
-        gplane=kwargs.get("gplane", "none"),
+    ps_backend.show_cad(
+        parts, part_options, labels, show_gui_panels=show_gui_panels, **kwargs
     )
-    ps_backend.add_features(labels, parts_list, part_options)
-
-    cam, target = _compute_default_camera(parts_list)
-    ps.look_at(cam, target)
-
-    # Polyscope's panel toggle is a process-wide global; set it explicitly
-    # both ways so a previous call with the opposite setting doesn't leak.
-    ps.set_build_default_gui_panels(show_gui_panels)
-
-    ps.show()
 
 
 __all__ = [
